% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{boxoffice}
\alias{boxoffice}
\title{Download Information on Box Office Results for Movies}
\usage{
boxoffice(dates = Sys.Date() - 1, site = "mojo", top_n = NULL)
}
\arguments{
\item{site}{Whether you want to get data from boxofficemojo.com or the-numbers.com.}
}
\value{
Data frame returning info on the name of the movie, its daily gross,
 gross-to-date, and gross-per-theater for each date inputted.
}
\description{
Download Information on Box Office Results for Movies
}
\examples{
boxoffice()
}
